Changed float to double, short to long int, int to long int types throughout in GEMIPM2K and in the respective parts of GEMS-PSI (with DODs for MULTI data objects); replaced outdated calls (like memset(), memcpy() etc.) with loops, all this to improve compatibility with parallel architectures (SD)
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@1121 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/gem2mt-sep/ms_gem2mt.cpp
@@ -413,7 +413,7 @@ int TGEM2MT::MassTransInit( const char *chbr_in1 )
   TNodeArray::na = na = new TNodeArray( /* mtp->xC,mtp->yC,mtp->zC */ mtp->nC );
 
  // Prepare the array for initial conditions allocation
-  int* nodeType = new int[mtp->nC];
+  long int* nodeType = new long int[mtp->nC];
   for( ii =0; ii<mtp->nC; ii++ )
          nodeType[ii] = mtp->DiCp[ii][0];
